+++ src/components/blocks/block-post-form/block-post-from.scss
@@ -1,0 +1,1 @@
+@import 'components/app/app';
